print_buf_lines - cast to unsigned to avoid warnings

+++ menu/menu_displaylist.c
@@ -13301,7 +13301,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
          case DISPLAYLIST_CORE_CONTENT:
             menu_entries_clear(info->list);
 #ifdef HAVE_NETWORKING
-            count = print_buf_lines(info->list, menu->core_buf, "",
+            count = (unsigned)print_buf_lines(info->list, menu->core_buf, "",
                   menu->core_len, FILE_TYPE_DOWNLOAD_CORE_CONTENT,
                   true);
 
@@ -13334,7 +13334,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
                   strlcpy(menu->core_buf, con, menu->core_len);
                free(info_path_cpy);
 
-               if ((count = print_buf_lines(
+               if ((count = (unsigned)print_buf_lines(
                            info->list, menu->core_buf, new_label,
                            menu->core_len, FILE_TYPE_DOWNLOAD_URL,
                            false)) == 0)
@@ -13364,7 +13364,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
                      network_buildbot_assets_url,
                      "cores", sizeof(new_label));
 
-               if ((count = print_buf_lines(info->list, menu->core_buf, new_label,
+               if ((count = (unsigned)print_buf_lines(info->list, menu->core_buf, new_label,
                            menu->core_len, FILE_TYPE_DOWNLOAD_URL, true)) == 0)
                   menu_entries_append(info->list,
                         msg_hash_to_str(MENU_ENUM_LABEL_VALUE_NO_ENTRIES_TO_DISPLAY),
@@ -13381,7 +13381,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
          case DISPLAYLIST_CORE_SYSTEM_FILES:
             menu_entries_clear(info->list);
 #ifdef HAVE_NETWORKING
-            count = print_buf_lines(info->list, menu->core_buf, "",
+            count = (unsigned)print_buf_lines(info->list, menu->core_buf, "",
                   menu->core_len, FILE_TYPE_DOWNLOAD_CORE_SYSTEM_FILES,
                   true);
 
@@ -13485,7 +13485,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
          case DISPLAYLIST_THUMBNAILS_UPDATER:
             menu_entries_clear(info->list);
 #ifdef HAVE_NETWORKING
-            count = print_buf_lines(info->list, menu->core_buf, "",
+            count = (unsigned)print_buf_lines(info->list, menu->core_buf, "",
                   menu->core_len, FILE_TYPE_DOWNLOAD_THUMBNAIL_CONTENT,
                   true);
 
@@ -13523,7 +13523,7 @@ bool menu_displaylist_ctl(enum menu_displaylist_ctl_state type,
          case DISPLAYLIST_LAKKA:
             menu_entries_clear(info->list);
 #ifdef HAVE_NETWORKING
-            count = print_buf_lines(info->list, menu->core_buf, "",
+            count = (unsigned)print_buf_lines(info->list, menu->core_buf, "",
                   menu->core_len, FILE_TYPE_DOWNLOAD_LAKKA,
                   true);
